Rename main.js to plete.js
