+++ rollup.config.js
@@ -2,34 +2,34 @@ import { terser } from "rollup-plugin-terser";
 
 export default [
   {
-    input: "lib/main.js",
+    input: "lib/plete.js",
     output: [
       {
-        file: "dist/main.js",
+        file: "dist/plete.js",
         format: "iife",
         name: "Plete"
       },
       {
-        file: "dist/main.min.js",
+        file: "dist/plete.min.js",
         format: "iife",
         name: "Plete",
         plugins: [terser()]
       },
       {
-        file: "dist/main-cjs.js",
+        file: "dist/plete-cjs.js",
         format: "cjs"
       },
       {
-        file: "dist/main-cjs.min.js",
+        file: "dist/plete-cjs.min.js",
         format: "cjs",
         plugins: [terser()]
       },
       {
-        file: "dist/main-esm.js",
+        file: "dist/plete-esm.js",
         format: "esm"
       },
       {
-        file: "dist/main-esm.min.js",
+        file: "dist/plete-esm.min.js",
         format: "esm",
         plugins: [terser()]
       }
